+++ corelib/src/util3d.cpp
@@ -465,8 +465,8 @@ std::multimap<int, pcl::PointXYZ> generateWords3DMono(
 				refCorners = refCornersRefined;
 				newCorners = newCornersRefined;
 
-				cv::Mat x(3, refCorners.size(), CV_64FC1);
-				cv::Mat xp(3, refCorners.size(), CV_64FC1);
+				cv::Mat x(3, (int)refCorners.size(), CV_64FC1);
+				cv::Mat xp(3, (int)refCorners.size(), CV_64FC1);
 				for(unsigned int i=0; i<refCorners.size(); ++i)
 				{
 					x.at<double>(0, i) = refCorners[i].x;
